Draft: Use size_t for length and index fields

+++ source/include/core_mqtt_serializer.h
@@ -151,7 +151,7 @@ typedef struct MQTTConnectInfo
     /**
      * @brief MQTT keep alive period.
      */
-    uint16_t keepAliveSeconds;
+    size_t keepAliveSeconds;
 
     /**
      * @brief MQTT client identifier. Must be unique per client.
@@ -161,7 +161,7 @@ typedef struct MQTTConnectInfo
     /**
      * @brief Length of the client identifier.
      */
-    uint16_t clientIdentifierLength;
+    size_t clientIdentifierLength;
 
     /**
      * @brief MQTT user name. Set to NULL if not used.
@@ -171,7 +171,7 @@ typedef struct MQTTConnectInfo
     /**
      * @brief Length of MQTT user name. Set to 0 if not used.
      */
-    uint16_t userNameLength;
+    size_t userNameLength;
 
     /**
      * @brief MQTT password. Set to NULL if not used.
@@ -181,7 +181,7 @@ typedef struct MQTTConnectInfo
     /**
      * @brief Length of MQTT password. Set to 0 if not used.
      */
-    uint16_t passwordLength;
+    size_t passwordLength;
 } MQTTConnectInfo_t;
 
 /**
@@ -203,7 +203,7 @@ typedef struct MQTTSubscribeInfo
     /**
      * @brief Length of subscription topic filter.
      */
-    uint16_t topicFilterLength;
+    size_t topicFilterLength;
 } MQTTSubscribeInfo_t;
 
 /**
@@ -235,7 +235,7 @@ typedef struct MQTTPublishInfo
     /**
      * @brief Length of topic name.
      */
-    uint16_t topicNameLength;
+    size_t topicNameLength;
 
     /**
      * @brief Message payload.
